+++ packages/studio-importor/src/app/import-schema/import-from-yaml/index.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { Space, Tooltip, Upload, Button, message, notification } from 'antd';
+import { Space, Tooltip, Upload, Button, message, notification, Typography, Flex } from 'antd';
 import type { UploadProps } from 'antd';
 import { InboxOutlined } from '@ant-design/icons';
 import yaml from 'js-yaml';
@@ -81,9 +81,15 @@ const ImportFromYAML = (props: IProps) => {
         multiple={true}
         style={{ height: '100px', width: '100%', ...style }}
       >
-        <div style={{ display: 'flex', flexDirection: 'column', justifyContent: 'center' }}>
-          <p className="ant-upload-drag-icon">{icon}</p>
-        </div>
+        <Flex justify="center" align="center" vertical gap={24}>
+          {icon}
+          <Typography.Text type="secondary">
+            For the definition and description of the schema model, please refer to the{' '}
+            <a href="https://graphscope.io/docs/flex/interactive/data_model" target="_blank">
+              document
+            </a>
+          </Typography.Text>
+        </Flex>
       </Dragger>
     </div>
   );