+++ examples/graphy-data/CHANGELOG.md
@@ -1,5 +1,17 @@
 # llm-workflow-graph
 
+## 0.1.2
+
+### Patch Changes
+
+- update graph
+- Updated dependencies
+  - @graphscope/studio-components@0.1.2
+  - @graphscope/studio-importor@0.1.2
+  - @graphscope/graph-apps@0.1.2
+  - @graphscope/studio-graph@0.1.2
+  - @graphscope/studio-query@0.1.2
+
 ## 0.1.1
 
 ### Patch Changes