+++ packages/studio-components/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @graphscope/studio-components
 
+## 0.1.2
+
+### Patch Changes
+
+- update graph
+
 ## 0.1.1
 
 ### Patch Changes